requirements.txt
python-binance==1.0.16
flask==2.0.1
werkzeug==2.0.3 